Source code/Server/src/requirements.txt
django

djangorestframework

numpy

mysqlclient

scipy

pandas

requests

beautifulsoup4

nltk

django-cors-headers

djangorestframework-jwt

pymysql
